Fix some timing issues in reading from BM70

+++ include/timer1.h
@@ -11,4 +11,4 @@ void delay1(unsigned long ms);
 /**
  * Read some bytes off a serial device with a timeout
  */
-int readBytes(HardwareSerial * serial, uint8_t * buffer, size_t length, uint8_t timeout_ms);
+size_t readBytes(HardwareSerial * serial, uint8_t * buffer, size_t length, int timeout_ms);
